# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Delivery API",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insira o token JWT no formato: Bearer {token}"
    }
  },
...
  "paths": {
...
    "/reviews/company/{companyId}": {
      "get": {
        "tags": [
          "Avaliações"
        ],
        "summary": "Listar avaliações públicas de uma empresa",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Avaliações da empresa"
          }
        }
      }
    },
...
    "/company-orders": {
      "get": {
        "tags": [
          "Pedidos da Empresa"
        ],
        "summary": "Listar pedidos recebidos pela empresa logada",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pedidos",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "integer"
                  },
                  "companyId": {
                    "type": "integer"
                  },
                  "addressId": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  },
                  "total": {
                    "type": "number"
                  },
                  "createdAt": {
                    "type": "string"
                  },
                  "nomeCliente": {
                    "type": "string",
                    "description": "Nome do usuário que fez o pedido"
                  },
                  "mediaAvaliacoes": {
                    "type": "number",
                    "description": "Média de avaliações que o usuário recebeu"
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado (não é empresa)"
          }
        }
      }
    },
    "/company-orders/{id}/assign": {
      "patch": {
        "tags": [
          "Pedidos da Empresa"
        ],
        "summary": "Atribuir pedido a um motoboy da empresa",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "atribuição",
            "schema": {
              "type": "object",
              "properties": {
                "motoboyId": {
                  "type": "string"
                }
              },
              "required": [
                "motoboyId"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Motoboy atribuído com sucesso"
          },
          "400": {
            "description": "motoboyId ausente"
          },
          "404": {
            "description": "Pedido ou motoboy não encontrado"
          }
        }
      }
    },
    "/company-orders/{id}/available-motoboys": {
      "get": {
        "tags": [
          "Pedidos da Empresa"
        ],
        "summary": "Listar motoboys ativos da empresa",
        "description": "Retorna todos os motoboys ativos da empresa logada para seleção",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do pedido (opcionalmente usado no front para indicar contexto)"
          }
        ],
        "responses": {
          "200": {
            "description": "Motoboys disponíveis",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "nome": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "telefone": {
                    "type": "string"
                  },
                  "placaVeiculo": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/company-orders/{id}/status": {
      "patch": {
        "tags": [
          "Pedidos da Empresa"
        ],
        "summary": "Atualizar status do pedido",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "status",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "enum": [
                    "pendente",
                    "aceito",
                    "em preparo",
                    "saiu para entrega",
                    "entregue",
                    "cancelado"
                  ]
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status atualizado com sucesso"
          },
          "404": {
            "description": "Pedido não encontrado"
          },
          "403": {
            "description": "Acesso negado"
          }
        }
      }
    },
...
    "/orders": {
      "post": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Criar novo pedido",
        "description": "Cria um novo pedido para a empresa especificada. Usuários bloqueados pela empresa não poderão realizar pedidos.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "pedido",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "companyId": {
                  "type": "integer",
                  "example": 123
                },
                "addressId": {
                  "type": "string",
                  "example": "end-456"
                },
                "items": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "productId": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      },
                      "price": {
                        "type": "number"
                      },
                      "quantity": {
                        "type": "integer"
                      },
                      "observation": {
                        "type": "string"
                      },
                      "addOns": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "type": "string"
                            },
                            "price": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    },
                    "required": [
                      "productId",
                      "name",
                      "price",
                      "quantity"
                    ]
                  }
                }
              },
              "required": [
                "companyId",
                "addressId",
                "items"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Pedido criado com sucesso"
          },
          "400": {
            "description": "Campos obrigatórios faltando"
          },
          "403": {
            "description": "Usuário bloqueado pela empresa"
          },
          "404": {
            "description": "Empresa não encontrada"
          }
        }
      },
      "get": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Listar pedidos do usuário logado",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pedidos"
          }
        }
      }
    },
...
    "/companies": {
      "get": {
        "tags": [
          "Empresas"
        ],
        "summary": "Listar todas as empresas",
        "responses": {
          "200": {
            "description": "Lista de empresas"
          }
        }
      },
      "post": {
        "tags": [
          "Empresas"
        ],
        "summary": "Criar nova empresa",
        "parameters": [
          {
            "in": "body",
            "name": "empresa",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "image": {
                  "type": "string"
                },
                "rating": {
                  "type": "number"
                },
                "deliveryFee": {
                  "type": "number"
                },
                "deliveryTime": {
                  "type": "string"
                },
                "categories": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "dishes": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              },
              "required": [
                "name",
                "image",
                "rating",
                "deliveryFee",
                "deliveryTime",
                "categories"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Empresa criada com sucesso"
          },
          "400": {
            "description": "Dados obrigatórios faltando"
          }
        }
      }
    },
    "/companies/{id}": {
      "get": {
        "tags": [
          "Empresas"
        ],
        "summary": "Buscar empresa por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Empresa encontrada"
          },
          "404": {
            "description": "Empresa não encontrada"
          }
        }
      },
      "patch": {
        "tags": [
          "Empresas"
        ],
        "summary": "Atualizar empresa por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "empresa",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "image": {
                  "type": "string"
                },
                "rating": {
                  "type": "number"
                },
                "deliveryFee": {
                  "type": "number"
                },
                "deliveryTime": {
                  "type": "string"
                },
                "categories": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "dishes": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Empresa atualizada"
          },
          "404": {
            "description": "Empresa não encontrada"
          }
        }
      },
      "delete": {
        "tags": [
          "Empresas"
        ],
        "summary": "Remover empresa por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Empresa removida com sucesso"
          },
          "404": {
            "description": "Empresa não encontrada"
          }
        }
      }
    },
...
  }
}